requirements.txt
certifi==2020.6.20
cycler==0.10.0
graphviz==0.14.2
kiwisolver==1.3.1
matplotlib==3.3.2
numpy==1.19.2
Pillow==8.0.1
pip==20.2.1
pydot==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.5.2
setuptools==49.2.1
six==1.15.0
